+++ test/plugin/parsed_json_test.rb
@@ -1,0 +1,14 @@
+require "test_helper"
+
+describe "the parsed_json plugin" do
+  before do
+    @attacher = attacher { plugin :parsed_json }
+  end
+
+  it "enables assigning cached files with hashes" do
+    cached_file = @attacher.cache.upload(fakeio)
+    @attacher.assign(cached_file.data)
+
+    assert_kind_of Shrine::UploadedFile, @attacher.get
+  end
+end